+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## ChangeLog
 
 * Added beam and trail emitters lifetime.
+* Added command to remove fx from block / entities.